#190: drop component interfaces — concrete PeerTransferComponent and TransferDetailsComponent
Per existing project canon in docs/engineering/presentation-layer.md §Component anatomy + §Screens and previews, Components are concrete classes; previews target stateless XxxContent(state, callbacks), not the Component. The interface-plus-Default split introduced earlier in this PR did not earn its place under that canon — preview seam works without it, and Decompose ChildStack typing already permits concrete components (RootComponent itself is concrete).

- Interfaces PeerTransferComponent and TransferDetailsComponent deleted.
- Former Default*Component classes renamed to PeerTransferComponent / TransferDetailsComponent (same FQNs as the deleted interfaces, no caller-side import churn).
- Test `onRetryAll delegates to peer component` deleted: relied on an anonymous interface impl as fake-peer; without interface there is no clean seam, and the asserted invariant — one-line delegate `details.onRetryAll() -> peer.onRetry()` — is exercised by the surrounding integration-style tests through real PeerTransferComponent.
- Stub class `presentation.TransferDetailsComponent` left as-is for now; RootComponent still references it. Integration with the full TransferDetailsComponent belongs to #191 where the navigation factory wiring lands.
