+++ composeApp/src/commonMain/kotlin/com/tubetoast/tether/presentation/transfer/TransferDetailsComponent.kt
@@ -1,17 +1,22 @@
 package com.tubetoast.tether.presentation.transfer
 
+import com.arkivanov.decompose.ComponentContext
 import com.arkivanov.decompose.value.Value
 
-interface TransferDetailsComponent {
-    val state: Value<PeerTransferState>
+class TransferDetailsComponent(
+    componentContext: ComponentContext,
+    private val peerComponent: PeerTransferComponent,
+    private val onBack: () -> Unit,
+) : ComponentContext by componentContext {
+    val state: Value<PeerTransferState> get() = peerComponent.state
 
-    fun onRetryFile(name: String)
+    fun onRetryFile(name: String) = peerComponent.onRetryFile(name)
 
-    fun onCancelFile(name: String)
+    fun onCancelFile(name: String) = peerComponent.onCancelFile(name)
 
-    fun onCancelTransfer()
+    fun onCancelTransfer() = peerComponent.onCancel()
 
-    fun onRetryAll()
+    fun onRetryAll() = peerComponent.onRetry()
 
-    fun onBack()
+    fun onBack() = onBack.invoke()
 }
